Use gevent sleep for switching context
`time.sleep` was changed to `gevent.sleep` to allow current greenlet to
sleep and others to run. When `time.sleep` is used, greenlet's context
is not changed from the main process to test's greenlet.

As a result of this, there is no data received by the main process while
hanging the tarantool server process and the test is fallen down by the
common timeout (NO_OUTPUT_TIMEOUT). Even worse, the tarantool server
process is not killed by test-run.

Using `gevent.sleep` makes the test fail by the test timeout and kill
the farantool server process.

Part of #276