Retry on a Lua error in test_run:wait_cond()
I think that a developer expects that wait_cond() will continue retrying
on any failure: as on a negative result from the condition function, as
well as when it raises a Lua error.

See motivating example in the linked issue: it was not obvious that a
condition function may raise an error in this case. I guess it is more
or less frequent case.

Fixes https://github.com/tarantool/tarantool-qa/issues/128