FIXUP: Don't recreate TestSuite instances
To be honest, I just afraid of possible side effects from the double
execution of the TestSuite initialization code. And it looks as a bad
pattern anyway.

I just moved the check to the TestSuite constructor.

The error message is weird now (the full backtrace), I will prettify it
in the next commit.