Decrease unix socket filename length
See https://github.com/tarantool/tarantool/pull/6365 for details about
possible problems and the description how it occurs.

We're going to use tags like 2.10.0-beta1 and so the build directory
path under rpmbuild will be even longer than before. So we should
decrease the absolute path length of a unix socket and so it worth to
trim the basename.

Of course, sooner or later we should elaborate a full solution.