luatest: prefer built-in luatest over system's one
The previous commit deliberately introduces a regression, which is fixed
here. The fix, in short, is to don't duplicate default PATH entries at
start of the new PATH value. However a background information is
necessary to properly describe the regression and the fix.

test-run is suitable for two a bit different cases:

* Testing of tarantool.
* Testing of a module.

In order to support the first case, test-run looks for `tarantool` and
`tarantoolctl` executables locally in tarantool's build directory.
However if local executables are not found, test-run assumes that we
want to test a module and uses system `tarantool` and `tarantoolctl`
executables[^1].

Aside of saving the directories internally, test-run also place them
into the `PATH` environment variable. I guess it is to ease usage of the
executables in tests.

Let's assume the second case: we test a module and test-run uses a
system tarantool. Let's also assume that `tarantool` and `tarantoolctl`
executables are in `/usr/bin`.

In given conditions test-run duplicates `/usr/bin` twice at start of a
`PATH` value. It is almost harmless. But not entirely: if `luatest`
executable is in `/usr/bin` as well, test-run will choose it instead of
the built-in one. In fact, it is the regression from the previous
commit.

To make it clear, it looks so:

```
PATH=/usr/bin:/usr/bin:/path/to/test-run/lib/luatest/bin:<...>
```

This commit fixes the regression: now test-run doesn't place a path into
`PATH` if it is already there. So the directory with built-in `luatest`
executable will always go before system ones:

```
PATH=/path/to/test-run/lib/luatest/bin:<...>
```

I also use the freshly introduced `prepend_path()` helper function in
`lib/__init.py`, but it is purely to make the code concise.

Part of #325

[^1]: This behaviour may be misleading if one forgot to build tarantool
      and run tarantool's testing. There is an issue (#249) to revisit
      it.