Add option to tune test run retries after failure
This change adds the --retries (-r) option for having a facility to tune
the number of test run retries after a failure. Besides regular tests,
it is also the default value for 'fragile' tests unless the `retries`
option is set in the suite.ini config file.

Also, it introduces the `TEST_RUN_RETRIES` env variable to tune retries
count via shell (useful for CI systems).

Closes #334