Free port auto resolving for AppServer
This change replaces manual choosing a free iproto port for AppServer
by the auto resolving mechanism. In two words, test-run always provides
'127.0.0.1:0' as a value for LISTEN env variable that is used in a lua
test script. In this way, the iproto port will be picked automatically,
but if the test needs the real value of the port, it has to execute
`box.info.listen` for tarantool version >= 2.4.1, or other lua code for
tarantool version < 2.4.1.

Part of #141