Fix detection of debug builds
The `tarantool -v` output is a single multiline string.
Flag re.MULTILINE is required to match it with start of line (`^`) or
end of line (`$`) characters. The `.*` in the middle of the pattern
does not match start/end of lines, so this expression can not match
something like this:

```
Target: Darwin-x86_64-Release
Build options: ... -Debug ...
```

The re.IGNORECASE is not required, because the build mode is always
spelled in one particular way: `-Debug`.

Follow-up to #352
Fixes #355

Co-authored-by: Serge Petrenko <sergepetrenko@tarantool.org>