Fix non-supported character in log file name
Running jobs in tarantool-ci fails at the action/upload-artifact runtime
if there are broken tests. The reason for the problem is that the path
of the test artifacts file contains the `:` character when test has been
parametrized.

Its non-supported character in the upload action. The action doesn't
allow the following characters in artifact path: ",:,<,>,|,*,?. [1]

To fix the problem, the `:` character has been replaced with `.`.

[1] https://github.com/actions/toolkit/blob/master/packages/artifact/docs/additional-information.md#non-supported-characters

Fixes tarantool/tarantool-qa#90