Adjust test result report width to terminal size
Now the test configuration name is not trimmed if the terminal width
allows it.

The idea can be illustrated this way:

```
+----------------------------------------------------------------------+
| Wide terminal                                                        |
+----------------------------------------------------------------------+
| suite/mytest_test.lua my_configuration                  [ pass ]     |
+----------------------------------------------------------------------+

+-------------------------------------------------+
| Narrow terminal                                 |
+-------------------------------------------------+
| suite/mytest_test.lua my_configur> [ pass ]     |
+-------------------------------------------------+
```

I want to add support for ability to run test cases as separate parallel
tasks for luatest based tests with reporting of the test case name as a
test configuration to the terminal. The test case names sometimes are
quite long and it is convenient to see as much symbols from it as
possible.

At the same time, the configuration name is stripped if the terminal has
a low width. This way the output look good as on a wide as well as on a
narrow terminal.

The feature works on Python 3.3+.