Fail if core=tarantool default server failed
Currently we stop default server in worker run loop only before the
next test start. So the previous test does not account whether the
Tarantool executing the test exits successfully or not. Thus we
do not fail on memory leaks for example.

Let's stop the server and account its exit status in test.

Closes #416