Remove luatest and checks submodules
The `test-run` project is rarely updated these days while `luatest` is
rapidly developing. Let's remove the `luatest` submodule from `test-run`
to ease luatest bumps in projects using `test-run`. Now, `luatest` is
supposed to be added as a submodule to all projects using `test-run`.

By default, `test-run` looks for `luatest` in `../third_party/luatest`,
but this can be overridden with the new `--luatestdir option`.

Note that we also remove the `checks` submodule, which was added as
a dependency of `luatest`. It was built in Tarantool in 2.11.0 so
there's no need in it anymore.

Closes #453